app/backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
pandas==2.1.0
numpy==1.25.2
torch==2.0.1
python-multipart==0.0.6
requests==2.31.0
pillow==10.0.1
python-dotenv==1.0.0
openai==1.57.2